Refactor loading functions to tools
Remvoed unused variable
improved annotations and loading count
added arguments for parallel jobs, verbosity
Ignore some more output files
reworked recreation file to be less verbose and with timings
Save more models for the lunar lander environment

+++ interpretable_ddts/tools.py
@@ -3,11 +3,13 @@ from __future__ import annotations
 import argparse
 from pathlib import Path
 import re
-from typing import Iterable, Union
+from typing import Iterable, Literal, Union, Optional, TYPE_CHECKING
 import pandas as pd
 import random
 import numpy as np
 import torch
+if TYPE_CHECKING:
+    from pandas._typing import AggFuncTypeBase
 
 RE_PARSE_FILENAME = re.compile(
     r"(?P<parent_dir>.+?/)?"  # likely for model files
@@ -29,30 +31,98 @@ RE_PARSE_FILENAME_OLD = re.compile(
     r"(?:_v(?P<version>\d+))?"  # no version in old format
 )
 
-def match_filename(filename: Union[str, Path]) -> "re.Match[str] | None":
-    if isinstance(filename, Path):
-        filename = filename.name
+def match_filename(filename: str) -> "re.Match[str] | None":
     result = RE_PARSE_FILENAME.match(filename)
     if result is None:
         result = RE_PARSE_FILENAME_OLD.match(filename)
     return result
 
-def seed_everything(env, seed):
-    random.seed(seed)
-    np.random.seed(seed)
-    # os.environ["PYTHONHASHSEED"] = str(seed)
-    if seed is None:
-        torch.seed()
-        torch.cuda.seed()
-    else:
-        torch.manual_seed(
-            seed
-        )  # setting torch manual seed causes bad models, # ok seed 124
-        torch.cuda.manual_seed_all(seed)
-    # torch.backends.cudnn.deterministic = True
-    if env:
-        env.seed(seed)
-        env.action_space.seed(seed)
+def parse_filename(filename: Union[str, Path]):
+    if isinstance(filename, Path):
+        filename = filename.name
+    filename = filename.split("/")[-1]
+    result = match_filename(filename)
+    assert result is not None, f"Filename {filename} does not match pattern"
+    data = result.groupdict()
+    data["version"] = int(data["version"])
+    data["num"] = int(data["num"])
+    del data["features"]
+    data["GPU"] = bool(data["GPU"])
+    data["typ"] = "hidden layers" if data["typ"] == "hid" else data["typ"]
+    return data
+
+
+def load_rewards(files: Iterable[Union[str, Path]]):
+    files = list(files)
+    headers = list(map(parse_filename, files))
+    print(files)
+    objs = (
+        pd.read_csv(file, header=None).T.set_index(
+            # create_df_index
+            pd.MultiIndex.from_tuples(
+                [
+                    (
+                        header["env"],
+                        header["method"],
+                        header["typ"],
+                        int(header["num"]),
+                        bool(header["GPU"]),
+                        int(header["version"]),
+                    )
+                ],
+                names=["env", "method", "sub-method", "capacity", "GPU", "version"],
+            )
+        )
+        for file, header in zip(files, headers)
+    )
+    data = pd.concat(objs)
+    data.sort_index(
+        inplace=True,
+    )
+    return data
+
+
+def load_output(
+    file: Union[str, Path],
+    index=("env", "method", "sub-method", "capacity", "GPU", "version"),
+    aggregate_version: Optional[Literal["max", "mean"] | AggFuncTypeBase] = None,
+    aggregate_column: str="discrete_reward",
+    **kwargs,
+):
+    df = pd.read_csv(file, index_col=index, **kwargs)
+    if not aggregate_version:
+        return df
+    df_2 = df.reset_index().set_index([*index, "episode"])
+    agg_df = (
+        df_2.groupby(list(index)).aggregate(
+            aggregate_version
+        ).sort_values(aggregate_column, ascending=False)
+    )
+    return agg_df
+    
+    
+def create_single_index(header: dict[str, str]):
+    return pd.MultiIndex(
+        (
+                header["env"],
+                header["method"],
+                header["typ"],
+                int(header["num"]),
+                bool(header["GPU"]),
+                int(header.get("version", 99) if header.get("version", 99) is not None else 99),  # no version in old format
+                int(header["episode"]),
+        ),
+         names=[
+            "env",
+            "method",
+            "sub-method",
+            "capacity",
+            "GPU",
+            "version",
+            "episode",
+        ],
+    )
+    
 
 def create_df_index(metadata: Iterable[dict[str, str]]):
     return pd.MultiIndex.from_tuples(
@@ -80,19 +150,22 @@ def create_df_index(metadata: Iterable[dict[str, str]]):
     )
 
 
-def parse_filename(filename: Union[str, Path]):
-    if isinstance(filename, Path):
-        filename = filename.name
-    filename = filename.split("/")[-1]
-    result = RE_PARSE_FILENAME.match(filename)
-    assert result is not None, f"Filename {filename} does not match pattern"
-    data = result.groupdict()
-    data["version"] = int(data["version"])
-    data["num"] = int(data["num"])
-    del data["features"]
-    data["GPU"] = bool(data["GPU"])
-    data["typ"] = "hidden layers" if data["typ"] == "hid" else data["typ"]
-    return data
+def seed_everything(env, seed, torch_manual=False):
+    random.seed(seed)
+    np.random.seed(seed)
+    # os.environ["PYTHONHASHSEED"] = str(seed)
+    if seed is None:
+        torch.seed()
+        torch.cuda.seed()
+    elif torch_manual:
+        torch.manual_seed(
+            seed
+        )  # setting torch manual seed causes bad models, # ok seed 124
+        torch.cuda.manual_seed_all(seed)
+    #
+    if env:
+        env.seed(seed)
+        env.action_space.seed(seed)
 
 
 if __name__ == "__main__":
@@ -104,6 +177,7 @@ if __name__ == "__main__":
         print("Cleaning up...")
         rewards_dir = Path("txts/")
         model_dir = Path("models/")
+        clean_files = []
         for file in rewards_dir.glob("*.txt"):
             txt = file.read_text().split("\n")
             if len(txt) < 1000:
@@ -112,7 +186,14 @@ if __name__ == "__main__":
                 models = list(model_dir.glob(f"*th{stem}*"))
                 if models:
                     for model in models:
-                        pass
-                        model.unlink()
+                        clean_files.append(model)
+                clean_files.append(file)
+        print("Files to remove:", clean_files)
+        do_clean = input("Do you want to remove these files? (y/n): ").lower()
+        if do_clean == "y":
+            for file in clean_files:
                 file.unlink()
+        else:
+            print("Aborted")
+
